requirements.txt
google-genai
dotenv
llama-index-llms-google-genai
llama-index
uvicorn
fastapi
pandas
tqdm

